+++ templates/index.html
@@ -21,8 +21,8 @@
         </h1>
         <!-- Add this line for the leaderboard link -->
         <div class="header-links">
-            <a href="{{ url_for('leaderboard') }}" class="leaderboard-link">排行榜</a>
-            <a href="{{ url_for('logout') }}">登出</a>
+            <a href="{{ url_for('leaderboard.leaderboard') }}" class="leaderboard-link">排行榜</a>
+            <a href="{{ url_for('auth.logout') }}">登出</a>
         </div>
     </div>
